use size_t type for iterating index in st matrix

+++ sparse_table.hpp
@@ -41,13 +41,13 @@ class sparse_table_matrix {
     public:
     sparse_table_matrix(const std::vector<std::vector<T>> &init, const Op &comb = {})
         : nx(init.size()), ny(init[0].size()), mx(std::bit_width(nx)), dp(mx), op(comb) {
-            for (int i = 0; i < nx; i++) {
+            for (size_t i = 0; i < nx; i++) {
                 dp[0].emplace_back(init[i], comb);
             }
-            for (int j = 1; j < mx; j++) {
-                for (int i = 0; i + (1 << j) <= nx; i++) {
+            for (size_t j = 1; j < mx; j++) {
+                for (size_t i = 0; i + (1 << j) <= nx; i++) {
                     std::vector<T> w(ny);
-                    for (int k = 0; k < ny; k++) {
+                    for (size_t k = 0; k < ny; k++) {
                         w[k] = op(dp[j - 1][i].query(k, k), dp[j - 1][i + (1 << (j - 1))].query(k, k));
                     }
                     dp[j].emplace_back(w, comb);
